requirements.txt
pyowm==3.3.0
matplotlib==3.6.2
mediapipe==0.9.1.0
onnxruntime==1.14.1
Pillow==9.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
pygame==2.4.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.9.1
imread-from-url==0.1.3
imutils==0.5.4
colormath==3.0.0
imread-from-url==0.1.3
googlemaps==4.7.3
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.0
tensorflow==2.10.0
pandas==1.4.4
firebase_admin == 6.2.0